Switch the back-end API to use a single list of Net::DNS::RRs
This isolates the structure decoding in Net::DNS::Create instead of making
all the backends also implement it. Now changes can be done in one place.

This simplifies the backends, except for Route53. Route53 groups records by
RR and name, so we have to do the same to the data. It's not too hard but
it's a complication the other backends dont share.